Update index.md
Getting started, roughing out the outline of what I am writing. Note: I relied heavy on the online installation, administration, and user guides for ownCloud in developing this guide, and, in the spirit of single-sourcing, re-used much of that content. Let's hear it for modules! :-) I also dipped into YouTube for walk-throughs and such, since it is always useful to see directly how something is done.